refactored to use os.path calls for basename and dirname, added dialog if try to compile bytecode

+++ main.py
@@ -368,7 +368,12 @@ class MainWindow(QMainWindow, Ui_MainWindow):
     def compile_files(self):
         local_file_paths = self.get_local_file_selection()
         for path in local_file_paths:
-            mpy_path = os.path.splitext(path)[0]+".mpy"
+            split = os.path.splitext(path)
+            if split[1] == ".mpy":
+               title = "COMPILE WARNING!! " + os.path.basename(path)
+               QMessageBox.warning(self, title, "Can't compile .mpy files, already bytecode")
+               continue
+            mpy_path = split[0]+".mpy"
             try:
                os.unlink(mpy_path)
             except:
@@ -376,10 +381,7 @@ class MainWindow(QMainWindow, Ui_MainWindow):
             self.localFilesTreeView.repaint()
             time.sleep(.500)
             QApplication.processEvents()   # not working to refresh before the compile...
-            last_slash_idx = path.rfind("/")+1
-            directory = path[:last_slash_idx]
-            name = path[last_slash_idx:]
-            subprocess.Popen([Settings().mpy_cross_path, name], cwd=directory)
+            subprocess.Popen([Settings().mpy_cross_path, os.path.basename(path)], cwd=os.path.dirname(path))
 
     def finished_read_mcu_file(self, file_name, transfer):
         assert isinstance(transfer, FileTransfer)
